# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для личного ассистента продуктивности UniFlow",
        "title": "UniFlow API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@uniflow.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:50031",
    "basePath": "/api",
    "paths": {
        "/auth/max": {
            "post": {
                "description": "Создает или получает пользователя по MAX User ID и возвращает JWT токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Аутентификация через MAX",
                "parameters": [
                    {
                        "description": "MAX User ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http_handlers.AuthWithMAXRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http_handlers.AuthWithMAXResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/contexts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех контекстов (учеба, проекты, личное) текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contexts"
                ],
                "summary": "Получить все контексты пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "contexts: array of Context objects",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новый контекст (например, \"Математика\", \"Курсовая работа\")",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contexts"
                ],
                "summary": "Создать новый контекст",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные контекста",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http_handlers.CreateContextRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_core_models.Context"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/contexts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает подробную информацию о контексте",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contexts"
                ],
                "summary": "Получить контекст по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_core_models.Context"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех задач текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить все задачи пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tasks: array of Task objects",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новую задачу с привязкой к контексту (опционально)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создать новую задачу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные задачи",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http_handlers.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_core_models.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/today": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список задач с дедлайном на сегодня",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить задачи на сегодня",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "tasks: array of Task objects",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_singl3focus_uniflow_internal_adapters_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_singl3focus_uniflow_internal_core_models.Context": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "Цвет для UI (HEX)",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deadlineAt": {
                    "description": "Опционально: дедлайн контекста",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "subjectID": {
                    "description": "Опционально: ID предмета в расписании",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_core_models.ContextType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "github_com_singl3focus_uniflow_internal_core_models.ContextType": {
            "type": "string",
            "enum": [
                "subject",
                "project",
                "personal",
                "work",
                "other"
            ],
            "x-enum-comments": {
                "ContextTypeOther": "Другое",
                "ContextTypePersonal": "Личное",
                "ContextTypeProject": "Проект",
                "ContextTypeSubject": "Учебный предмет",
                "ContextTypeWork": "Работа"
            },
            "x-enum-descriptions": [
                "Учебный предмет",
                "Проект",
                "Личное",
                "Работа",
                "Другое"
            ],
            "x-enum-varnames": [
                "ContextTypeSubject",
                "ContextTypeProject",
                "ContextTypePersonal",
                "ContextTypeWork",
                "ContextTypeOther"
            ]
        },
        "github_com_singl3focus_uniflow_internal_core_models.Task": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "contextID": {
                    "description": "Опционально: привязка к контексту",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dueAt": {
                    "description": "Опционально: дедлайн задачи",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_singl3focus_uniflow_internal_core_models.TaskStatus"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "github_com_singl3focus_uniflow_internal_core_models.TaskStatus": {
            "type": "string",
            "enum": [
                "todo",
                "in_progress",
                "completed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "TaskStatusTodo",
                "TaskStatusInProgress",
                "TaskStatusCompleted",
                "TaskStatusCancelled"
            ]
        },
        "internal_adapters_http_handlers.AuthWithMAXRequest": {
            "type": "object",
            "properties": {
                "max_user_id": {
                    "type": "string"
                }
            }
        },
        "internal_adapters_http_handlers.AuthWithMAXResponse": {
            "type": "object",
            "properties": {
                "max_user_id": {
                    "type": "string"
                },
                "token": {
                    "description": "JWT token для дальнейшей аутентификации",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_adapters_http_handlers.CreateContextRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "deadline_at": {
                    "description": "ISO 8601 format",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "internal_adapters_http_handlers.CreateTaskRequest": {
            "type": "object",
            "properties": {
                "context_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_at": {
                    "description": "ISO 8601 format",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}